Calculates the log likelihood function
This routine delives a numerical estimate of the log likelihood function. However, it relies on the classical UT of the Kalman filter, that is, it approximates the density with the density of the normal distribution. The method in CHS(2010) uses a mixture of normal distributions, so that needs to be addressed next