Linear anchor
The function has been modified in two meaningful ways. First, it now replaced the options in argument f, allowing for no anchor (what used to be called CES) and linear anchor. Second, the linear anchor has been coded,and it takes an extra argument for the factor loadings for the fourth measurement. See Appendix A7.1 in CHS(2010, Appendix)